test: move tests to proper class due to interface change
